backend/requirements.txt
django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
pymongo>=4.6
google-generativeai>=0.5
openai>=1.0
python-dotenv>=1.0
gunicorn>=22.0
